Changed colors and added partial windows support
Windows support seems imcomplete, but colors work. I suspect the culprit
is shutil not grabbing the terminal size correctly, and thus falling
back to the default 80 width, which is 1 wider than the default terminal
size. I'm not going to bother to fix this right now.

Colors were changed to the "standard" shockwave colorscheme, which is
0 = black
A = white
B = green

Signed-off-by: ZweiSpeedruns <zcm727@gmail.com>